+++ app/Livewire/User/RoleIndex.php
@@ -2,43 +2,62 @@
 
 namespace App\Livewire\User;
 
+use App\Exports\RoleExport;
+use App\Imports\RoleImport;
+use Illuminate\Support\Facades\Log;
 use Livewire\Component;
+use Livewire\WithFileUploads;
+use Livewire\WithPagination;
+use Maatwebsite\Excel\Facades\Excel;
+use RealRashid\SweetAlert\Facades\Alert;
 use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role;
 
 class RoleIndex extends Component
 {
-    public $id, $name, $roles;
+    use WithFileUploads;
+    use WithPagination;
+
+    public $roleId, $name;
     public $permissions = [];
     public $selectedPermissions = [];
     public $form = false;
     public $search = '';
+    public $formImport = 0;
+    public $fileImport;
+
     public function store()
     {
         $this->validate([
             'name' => 'required',
         ]);
         $role = Role::updateOrCreate(
-            ['id' => $this->id],
+            ['id' => $this->roleId],
             ['name' => $this->name],
         );
-        $role->syncPermissions();
         $role->syncPermissions($this->selectedPermissions);
+        Log::notice(auth()->user()->name . ' menyimpan data role ' . $role->name);
         flash('Role ' . $role->name . ' saved successfully.', 'success');
         $this->closeForm();
     }
     public function destroy($id)
     {
         $role = Role::find($id);
-        $role->delete();
-        flash('Role ' . $role->name . ' deleted successfully.', 'success');
+        if ($role->users()->count()) {
+            Log::notice(auth()->user()->name . ' gagal menghapus data role ' . $role->name);
+            flash('Role tidak bisa dihapus karena sedang dipakai', 'danger');
+        } else {
+            $role->delete();
+            Log::notice(auth()->user()->name . ' menghapus data role ' . $role->name);
+            flash('Role ' . $role->name . ' deleted successfully.', 'success');
+        }
     }
     public function edit($id)
     {
         $this->form = true;
         $role = Role::find($id);
         $this->name = $role->name;
-        $this->id = $role->id;
+        $this->roleId = $role->id;
         $this->permissions = Permission::pluck('name', 'id');
         $this->selectedPermissions = $role->permissions()->pluck('name');
     }
@@ -46,31 +65,60 @@ class RoleIndex extends Component
     {
         $this->form = true;
         $this->name = '';
-        $this->id = '';
+        $this->roleId = '';
         $this->permissions = Permission::pluck('name', 'id');
+        $this->selectedPermissions = [];
     }
     public function closeForm()
     {
         $this->form = false;
         $this->name = '';
-        $this->id = '';
+        $this->roleId = '';
         $this->selectedPermissions = [];
     }
-    public function placeholder()
+    public function export()
     {
-        return view('components.placeholder.placeholder-text');
+        try {
+            $time = now()->format('Y-m-d');
+            Log::notice( auth()->user()->name . ' mengekspor data role');
+            flash('Export successfully', 'success');
+            return Excel::download(new RoleExport, 'role_backup_' . $time . '.xlsx');
+        } catch (\Throwable $th) {
+            flash('Mohon maaf ' . $th->getMessage(), 'danger');
+        }
+    }
+    public function openFormImport()
+    {
+        $this->formImport = $this->formImport ?  0 : 1;
+    }
+    public function import()
+    {
+        try {
+            $this->validate([
+                'fileImport' => 'required|mimes:xlsx'
+            ]);
+
+            Excel::import(new RoleImport, $this->fileImport->getRealPath());
+            Log::notice(auth()->user()->name . ' mengimpor data role');
+            Alert::success('Success', 'Imported successfully');
+            return redirect()->route('role-permission');
+        } catch (\Throwable $th) {
+            flash('Mohon maaf ' . $th->getMessage(), 'danger');
+        }
     }
     public function mount()
     {
-
+        $this->permissions = Permission::pluck('name', 'id');
     }
     public function render()
     {
         $search = '%' . $this->search . '%';
-        $this->roles = Role::orderBy('name', 'asc')->with(['permissions'])
+        $roles = Role::with('permissions')
+            ->withCount('users')
             ->where('name', 'like', $search)
-            ->get();
-        $this->permissions = Permission::pluck('name', 'id');
-        return view('livewire.user.role-index');
+            ->orderBy('name', 'asc')
+            ->paginate(10);
+
+        return view('livewire.user.role-index', compact('roles'));
     }
 }
